+++ docs/news.rst
@@ -3,6 +3,14 @@ Release notes
 
 .. changelog
 
+Unreleased
+----------
+
+Added
+~~~~~
+
+- Add DEBUG-level messages to the :ref:`schedule.json` and :ref:`cancel.json` webservices.
+
 1.6.0 (2025-07-22)
 ------------------